backend/requirements.txt
# EchoSheild Backend Requirements

# Core
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6

# Data validation
pydantic==2.5.0

# Optional: LLM Integration
langchain==0.1.0
openai==1.3.0

# Optional: Additional utilities
python-dotenv==1.0.0
aiohttp==3.9.1
